front/requirements.txt
streamlit
requests
python-dotenv
pillow